requirements.txt
faiss-cpu>=1.8.0
numpy>=1.26.0
pypdf>=5.0.0
openai>=1.0.0
python-dotenv>=1.0.0
